refactor(common): combine utils and common modules
