+++ kipher-digest/build.gradle.kts
@@ -4,7 +4,6 @@ plugins {
 
 dependencies {
     api(projects.kipherCommon)
-    api(projects.kipherUtils)
     implementation(libs.bouncycastle.provider)
     implementation(libs.jetbrains.annotations)
 }
