+++ kipher-mac/build.gradle.kts
@@ -9,7 +9,6 @@ plugins {
 
 dependencies {
     api(projects.kipherCommon)
-    api(projects.kipherUtils)
     implementation(libs.bouncycastle.provider)
     implementation(libs.jetbrains.annotations)
 }
