+++ kipher-mac/src/main/kotlin/io/github/jhdcruz/kipher/mac/Hmac.kt
@@ -5,9 +5,9 @@
 
 package io.github.jhdcruz.kipher.mac
 
+import io.github.jhdcruz.kipher.common.Format.toHexString
 import io.github.jhdcruz.kipher.common.KipherException
 import io.github.jhdcruz.kipher.common.KipherProvider
-import io.github.jhdcruz.kipher.utils.Formatters.toHexString
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.annotations.Nullable
 import java.security.InvalidKeyException
