--- kipher-utils/README.md
@@ -1,9 +1,0 @@
-# Module kipher-utils
-
-Utility functions for kipher library.
-
-> **Note**
->
-> Consumers of this library should not use this module directly.
->
-> They are already provided by each `kipher` module.
