--- kipher-utils/gradle.properties
@@ -1,10 +1,0 @@
-#
-# Copyright 2023 Joshua Hero Dela Cruz
-# SPDX-License-Identifier: Apache-2.0
-#
-
-ARTIFACT_ID=kipher-utils
-VERSION=0.1.0-SNAPSHOT
-
-POM_NAME=kipher-utils
-POM_DESCRIPTION=Utility functions for kipher library.
